training/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.15.0
onnxruntime>=1.17.0
onnxscript>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
numpy>=1.26.0
pillow>=10.2.0
